+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Stores, where you are a jumper are now displayed at the bottom of the stores dropdown menu !3364 @AntonBallmaier
 - Improved the calendar export !3378 @AntonBallmaier
 - Pickups are displayed on dashboard for 30 minutes longer !3382
+- Removed address on businessCard !3362 @chriswalg
 
 ## Bugfixes
 - Changed the headline for group applications #1916 !3353 @ToLei